Revert breaking changes
